// front-end/src/app/shared/model/immobile.model.ts
export class Immobile {
    id: number;
    title: string;
    description: string;
    amount: number;
    area: number;
    address: string;
    addressNumber: string;
    complement: string;
    district: string;
    city: string;
    state: string;
    zipCode: string;
    created_at: string;
    updated_at: string;

}